pyproject.toml
[tool.poetry]
name = "education-test-task"
version = "0.1.0"
description = ""
authors = ["zhukovny <cruvar.zhukov@yandex.ru>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.1"
pillow = "^9.5.0"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
django-filter = "^23.2"
drf-spectacular = "^0.26.2"
black = "^23.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
